Modified the UserDefinedSQLType so that it uses Integer for nullability, type, and Boolean for autoIncrement, so that they can be nullable, which signifies to get the value from the upstream type. This follows the same 'inheritance' mechanism that the other properties follow, which allows for changes made in the upstream type to affect the non-overridden properties.
It is worth noting that in order to work with BeanUtils, the isAutoIncrement had to be changed to getAutoIncrement. Otherwise, BeanUtils will say that it cannot find a getter for autoIncrement, resulting in test failures.

There is a JIRA issue here about that: https://issues.apache.org/jira/browse/BEANUTILS-321

The argument for the current behaviour is that since Boolean isn't truly a boolean value (since it can be null), it should not use the isXXX format for getters.

+++ src/test/java/ca/sqlpower/sql/PLDotIniTest.java
@@ -290,7 +290,7 @@ public class PLDotIniTest extends TestCase {
     	assertEquals("VARCHAR", sqlType.getName());
     	assertEquals(BasicSQLType.TEXT, sqlType.getBasicType());
     	assertEquals("Test Type", sqlType.getDescription());
-    	assertEquals(12, sqlType.getType());
+    	assertEquals(12, (int) sqlType.getType());
     	
     	UserDefinedSQLType sqlTypeInList = target.getSQLTypes().get(0);
     	assertEquals(sqlType, sqlTypeInList);
